requirements.txt
langchain_google_genai
langgraph
langchain_community
langchain
accelerate
gradio
numpy
transformers
trimesh
grandalf
google
pillow